fix(CAT-21): add try catch and helper in api
