+++ packages/client/src/api/user.ts
@@ -1,18 +1,17 @@
+import { fetchHelper } from '@/helpers';
 import { YANDEX_API_HOST } from '@/helpers/constants/api';
 import { DataChangePassword } from '@/helpers/types/user';
 
 export class UserApi {
-  static async changePassword(passwordData: DataChangePassword) {
-    return await fetch(`${YANDEX_API_HOST}/user/password`, {
+  static async changePassword(params: DataChangePassword) {
+    return await fetchHelper({
+      url: '/user/password',
       method: 'PUT',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      credentials: 'include',
-      body: JSON.stringify(passwordData),
+      data: params,
     });
   }
 
+  //здесь что бы применить fetchHelper нужно его поменять, что бы была возможность менять Content-Type у headers
   static async changeAvatar(data: FormData) {
     return await fetch(`${YANDEX_API_HOST}/user/profile/avatar`, {
       method: 'PUT',
