+++ packages/client/src/pages/profile-page/profile-page.tsx
@@ -1,6 +1,7 @@
 import { PageWrapper, ProfileAvatar, ProfileForm } from '@/components';
 import './profile-page.css';
 import { useNavigate } from 'react-router-dom';
+import { ExitButton } from '@/components/exit-button';
 
 export const ProfilePage = () => {
   const navigate = useNavigate();
@@ -15,6 +16,9 @@ export const ProfilePage = () => {
           onClick={() => navigate('../change-password')}>
           Сменить пароль
         </p>
+        <div className='profile-page__exit'>
+          <ExitButton />
+        </div>
       </div>
     </PageWrapper>
   );
